requirements.txt

streamlit
Pillow
